Add various cantake functions
